Automatic and Repeating Reconnection to Server
I added a connection-status controlled while-loop in place of the singular connection request. If there is an error/exception thrown such as timeout or no hostname, or anything else, it posts that to the log (maybe change that for debug on only) and then waits as long as the user input on the preferences page, then tries again. It does this until it does not get an exception, therefore it breaks out of the while loop.